requirements.txt
streamlit>=1.30.0
openai>=1.10.0
chromadb>=0.4.22
pytrends>=4.9.2
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0
ddgs>=6.0.0
tiktoken>=0.5.0
pyyaml>=6.0
pytest>=7.0.0
